--- app/Traits/Retireable.php
@@ -1,81 +1,0 @@
-<?php
-
-namespace App\Traits;
-
-use App\Models\Retirement;
-
-trait Retireable
-{
-    /**
-     * Get the retirements of the model.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function retirements()
-    {
-        return $this->morphMany(Retirement::class, 'retiree');
-    }
-
-    /**
-     * Get the current retirement of the model.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function retirement()
-    {
-        return $this->morphOne(Retirement::class, 'retiree')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the previous retirement of the model.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function previousRetirement()
-    {
-        return $this->morphOne(Retirement::class, 'retiree')->whereNotNull('ended_at')->latest('started_at');
-    }
-
-    /**
-     * Retire the model.
-     *
-     * @return void
-     */
-    public function retire()
-    {
-        $this->retirements()->create(['started_at' => today()]);
-    }
-
-    /**
-     * Check to see if the model is retired.
-     *
-     * @return bool
-     */
-    public function isRetired()
-    {
-        return $this->retirements()->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Unretire the retired model.
-     *
-     * @return void
-     */
-    public function unretire()
-    {
-        $this->retirement()->update(['ended_at' => today()]);
-    }
-
-    /**
-     * Scope a query to only include retired models.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeRetired($query)
-    {
-        return $query->whereHas('retirements', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-}
